CH: Adding current location to search input
